requirements.txt
langgraph
langchain
langchain-groq
langchain-community
langchain-huggingface
langchain-pinecone
pinecone-client
tavily-python
python-dotenv
chainlit
pandas
pypdf
sqlalchemy
aiosqlite
sentence-transformers